requirements.txt
pandas
numpy
ccxt
feedparser
pyyaml
